requirements-test.txt
tox
coverage
path.py
easyprocess